DTIR-main/requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alibabacloud-oss-v2==1.1.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
aqtp==0.8.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.6.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.3.0
av==14.1.0
babel==2.16.0
beautifulsoup4==4.12.3
bitarray==3.0.0
bitsandbytes==0.45.0
bleach==6.2.0
blobfile==3.0.0
Brotli
certifi
cffi==1.17.1
chardet==5.2.0
charset-normalizer
chex==0.1.89
ciou==0.6.1
click==8.1.8
clip
cloudpickle==3.1.1
clu==0.0.12
comm==0.2.2
contourpy==1.3.1
crcmod==1.7
crcmod-plus==2.1.0
cryptography==44.0.3
cycler==0.12.1
Cython==3.0.11
datasets==3.5.0
debugpy==1.8.11
decorator==5.1.1
decord==0.6.0
deepspeed==0.16.3
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distrax==0.1.5
distro==1.9.0
dm-tree==0.1.9
docstring_parser==0.16
editdistance==0.8.1
einops==0.6.1
einops-exts==0.0.4
etils==1.12.0
eva-decord==0.6.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock
fire==0.7.0
flake8==7.1.1
flash_attn==2.7.4.post1
flatbuffers==25.2.10
flax==0.10.4
flaxformer
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
gast==0.6.0
gdown==5.2.0
gmpy2
google==3.0.0
google-pasta==0.2.0
gradio==4.29.0
gradio_client==0.16.1
gradio_imageslider==0.0.20
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
humanize==4.12.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
immutabledict==4.2.1
importlib_resources==6.5.2
iopath==0.1.10
iou==0.1.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==4.3.21
jax==0.5.2
jax-cuda12-pjrt==0.5.1
jax-cuda12-plugin==0.5.1
jaxlib==0.5.1
jedi==0.19.2
Jinja2
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.8
latex2mathml==3.77.0
lazy_loader==0.4
libclang==18.1.1
llama_models==0.1.1
llama_stack==0.1.1
llama_stack_client==0.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.0
mkl-service==2.4.0
mkl_fft
mkl_random
ml-dtypes==0.4.1
ml_collections==1.0.0
mpmath
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
namex==0.0.8
narwhals==1.20.1
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx
ninja==1.11.1.3
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.8.93
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open_clip_torch==2.32.0
openai==1.66.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optax
optree==0.14.0
orbax-checkpoint==0.11.6
orjson==3.10.13
oss2==2.19.1
ott-jax==0.2.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
panopticapi
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
portalocker==3.1.1
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.48
propcache==0.3.1
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyaml==25.1.0
pyarrow==19.0.1
pycocoevalcap==1.2
pycocotools==2.0.8
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.9.9
pycryptodomex==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.2.1
PySocks
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2020.5
PyYAML
pyzmq==26.2.0
qwen-vl-utils==0.0.8
referencing==0.35.1
regex==2024.11.6
requests
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
ruff==0.9.2
safetensors==0.5.0
scenic
scikit-image==0.25.2
scikit-learn==1.2.2
scipy==1.15.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
simple-parsing==0.1.7
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
svgwrite==1.4.3
sympy==1.13.1
tb-nightly==2.19.0a20250120
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-gan==2.1.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-probability==0.25.0
tensorflow-text==2.18.1
tensorflow_cpu==2.18.0
tensorstore==0.1.72
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.18.0
tfds-nightly==4.9.7.dev202502270044
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.8.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
toolz==1.0.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchreid==0.2.5
torchvision==0.21.0+cu124
tornado==6.4.2
tos==2.8.3
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
treescope==0.1.9
triton==3.2.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions
tzdata==2024.2
ultralytics==8.3.75
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3
uvicorn==0.34.0
volcengine==1.0.184
volcengine-python-sdk==2.0.2
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
